Extract Booter::ComponentRegistry into a separate file
